+++ src/MethodCall.php
@@ -340,7 +340,7 @@ class MethodCall extends Call
                     }
 
                     $expectedNumberParameters = $classMethod->getNumberOfRequiredParameters();
-                    if (!$expectedNumberParameters && $callNumberParameters > 0) {
+                    if (!$expectedNumberParameters && $callNumberParameters > 0 && !$classMethod->isVariadic()) {
                         $numberParameters = $classMethod->getNumberOfParameters();
                         if ($callNumberParameters > $numberParameters) {
                             throw new CompilerException(
@@ -469,7 +469,7 @@ class MethodCall extends Call
                                     $classMethod              = $classDefinition->getMethod($methodName);
                                     $expectedNumberParameters = $classMethod->getNumberOfRequiredParameters();
 
-                                    if (!$expectedNumberParameters && $callNumberParameters > 0) {
+                                    if (!$expectedNumberParameters && $callNumberParameters > 0 && !$classMethod->isVariadic()) {
                                         $numberParameters = $classMethod->getNumberOfParameters();
                                         if ($callNumberParameters > $numberParameters) {
                                             $className = $classDefinition->getCompleteName();